+++ metadetect/lsst/util.py
@@ -1299,3 +1299,36 @@ class MissingDataError(Exception):
 
     def __str__(self):
         return repr(self.value)
+
+
+def extract_shear_band_indices(mbexp, config):
+    """
+    Extract the indices requested shear bands, ensuring that each requested
+    band is available
+
+    Parameters
+    ----------
+    mbexp: MultiBandsExposure
+        The image data.  Must have .bands attribute
+    config: dict like
+        Must have entry 'shear_bands' holding band names
+
+    Returns
+    -------
+    List of band indices into the bands of mbexp
+    """
+    shear_band_names = config["shear_bands"] or mbexp.bands
+
+    if not all([sb in mbexp.bands for sb in shear_band_names]):
+        raise RuntimeError(
+            "Not all requested bands for shear are available. "
+            f"Bands `{shear_band_names}` were requested but the only "
+            f"bands available are `{mbexp.bands}`."
+        )
+
+    shear_band_indices = [
+        i for i, band in enumerate(mbexp.bands)
+        if band in shear_band_names
+    ]
+
+    return shear_band_indices